feat(mm-next): add `comscore-script` and import it in wholesite-script
