+++ packages/mirror-media-next/components/whole-site-script.js
@@ -1,7 +1,12 @@
 import AvividScript from './ads/avivid/avivid-script'
 import { useDisplayAd } from '../hooks/useDisplayAd'
-
+import ComScoreScript from './comscore-script'
 export default function WholeSiteScript() {
   const shouldShowAd = useDisplayAd()
-  return <>{shouldShowAd && <AvividScript />}</>
+  return (
+    <>
+      {shouldShowAd && <AvividScript />}
+      <ComScoreScript />
+    </>
+  )
 }
